+++ packages/bcgov-theme/stories/callout/Callout.stories.mdx
@@ -1,0 +1,28 @@
+import { Meta, Story, Canvas } from '@storybook/addon-docs/blocks';
+import Callout from '../../src/Callout';
+import * as stories from './Callout.stories.tsx';
+import { ArgsTable } from '@storybook/addon-docs/blocks';
+import { argTypes } from './args';
+
+<Meta title="Components/Callout" component={Callout} argTypes={argTypes} />
+
+# Callout
+
+Callouts are an excerpt of text that has been pulled out and used as a visual clue to draw the eye to the text. They are used to help direct a user's attention to important pieces of information.
+The callout component can either be passed content through the "content" prop, or passed children to render.
+
+export const Template = (args) => <Callout {...args} />
+
+<Canvas>
+  <Story story={stories.Default} name="Default" />
+</Canvas>
+
+## Arguments
+
+<ArgsTable story="Default" />
+
+## HTML only
+
+<Canvas>
+  <Story story={stories.HTML} name="html" />
+</Canvas>
