+++ packages/bcgov-theme/stories/callout/args.ts
@@ -1,0 +1,15 @@
+export const argTypes = {
+  size: {
+    control: {
+      type: 'select',
+      options: ['small', 'medium', 'large'],
+    },
+    description: 'The size of the checkbox',
+  },
+  content: {
+    control: {
+      type: 'text',
+    },
+    description: 'String content to show in the callout. To render more complex content, pass them as children.',
+  },
+};
